updated example unit test
